refactor: base module
