+++ test/e2e/base/base.go
@@ -3,18 +3,19 @@ package base
 import (
 	"time"
 
-	clientv3 "go.etcd.io/etcd/client/v3"
+	"github.com/go-resty/resty/v2"
 )
 
-// NewETCDClient create an ETCD client for testing.
-func NewETCDClient() (*clientv3.Client, error) {
-	cli, err := clientv3.New(clientv3.Config{
-		Endpoints:   []string{"127.0.0.1:12379"},
-		DialTimeout: 5 * time.Second,
-	})
-	if err != nil {
-		return nil, err
-	}
+const (
+	DefaultAddress = "127.0.0.1:12379"
+)
 
-	return cli, nil
+var (
+	httpClient = resty.New().SetTimeout(time.Second)
+)
+
+// GetHTTPClient returns the default HTTP client
+// timeout: 1 second
+func GetHTTPClient() *resty.Client {
+	return httpClient
 }
